tools/capsules_stat/pyproject.toml
[tool.poetry]
name = "capsules_stat"
version = "0.1.3"
description = ""
authors = ["Valeriy Zamaraiev <valeriy.zamaraiev@dfinity.org>"]

[tool.poetry.dependencies]
python = "^3.10"
python-gitlab = "^2.10.0"
GitPython = "^3.1.20"
absl-py = "^0.13.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
capsules_stat = "capsules_stat.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
